+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetMetricsInstrumentationOptions.cs
@@ -16,12 +16,12 @@ public sealed class AspNetMetricsInstrumentationOptions
     /// </summary>
     /// <param name="context"><see cref="HttpContext"/>: the HttpContext object. Both Request and Response are available.</param>
     /// <param name="tags"><see cref="TagList"/>: List of current tags. You can add additional tags to this list. </param>
-    public delegate void EnrichFunc(HttpContext context, ref TagList tags);
+    public delegate void EnrichWithHttpContextAction(HttpContext context, ref TagList tags);
 
     /// <summary>
-    /// Gets or sets a function to enrich a recorded metric with additional custom tags.
+    /// Gets or sets a delegate to enrich a recorded metric with additional custom tags.
     /// </summary>
-    public EnrichFunc? Enrich { get; set; }
+    public EnrichWithHttpContextAction? EnrichWithHttpContext { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether emit `server.address` and `server.port` attributes.